+++ test/test_computations/test_cbrng.py
@@ -3,119 +3,169 @@ import numpy
 import pytest
 
 from helpers import *
-from test_computations.cbrng_ref import philox, threefry
+from test_computations.cbrng_ref import philox as philox_ref
+from test_computations.cbrng_ref import threefry as threefry_ref
 
+from reikna.core import Type
 from reikna.helpers import product
-from reikna.cbrng import CBRNG, create_counters, create_key
+from reikna.cbrng import CBRNG
+from reikna.cbrng.bijections import threefry, philox
+from reikna.cbrng.tools import KeyGenerator
+from reikna.cbrng.samplers import uniform_integer, uniform_float, normal_bm, gamma
 import reikna.cluda.dtypes as dtypes
+
+
+class TestBijection:
+
+    def __init__(self, name, words, bitness):
+        rounds = 20 if name == 'threefry' else 10
+        if name == 'philox':
+            bijection_func = philox
+        else:
+            bijection_func = threefry
+        self._name = name
+        self._words = words
+        self._bitness = bitness
+        self._rounds = rounds
+        self.bijection = bijection_func(bitness, words, rounds=rounds)
+
+        func = philox_ref if name == 'philox' else threefry_ref
+        self._reference_func = lambda ctr, key: func(bitness, words, ctr, key, Nrounds=rounds)
+
+    def reference(self, counters, keygen):
+        result = numpy.empty_like(counters)
+        for i in range(counters.shape[0]):
+            result[i] = self._reference_func(counters[i], keygen(i))
+        return result
+
+    def __str__(self):
+        return "{name}-{words}x{bitness}-{rounds}".format(
+            name=self._name, words=self._words, bitness=self._bitness, rounds=self._rounds)
 
 
 def pytest_generate_tests(metafunc):
 
-    if 'name_and_params' in metafunc.funcargnames:
+    if 'test_bijection' in metafunc.funcargnames:
 
         vals = []
         ids = []
 
         for name, words, bitness in itertools.product(['threefry', 'philox'], [2, 4], [32, 64]):
-            # This particular set is not supported by CBRNG,
-            # because the key in this case is only 32bit and cannot hold
-            # both seed and thread id.
-            if name == 'philox' and words == 2 and bitness == 32:
-                continue
+            val = TestBijection(name, words, bitness)
+            vals.append(val)
+            ids.append(str(val))
 
-            rounds = 20 if name == 'threefry' else 10
-
-            params = dict(words=words, bitness=bitness, rounds=rounds)
-            vals.append((name, params))
-            ids.append("{name}-{words}x{bitness}-{rounds}".format(name=name, **params))
-
-        metafunc.parametrize('name_and_params', vals, ids=ids)
+        metafunc.parametrize('test_bijection', vals, ids=ids)
 
 
-def rng_ref(ctr, size, name, seed, **params):
-    dtype = numpy.uint32 if params['bitness'] == 32 else numpy.uint64
-    result = numpy.empty((size, params['words']), dtype)
-    base_key = create_key(name, params, seed=seed)
+def test_kernel_bijection(thr, test_bijection):
 
-    func = philox if name == 'philox' else threefry
-
-    counter = numpy.zeros(params['words'], dtype)
-    counter[-1] += ctr
-
-    for i in range(size):
-        key = base_key.copy()
-        key[-1] += numpy.cast[key.dtype](i)
-
-        result[i] = func(params['bitness'], params['words'], counter, key, params['rounds'])
-
-    return result.T
-
-
-def test_raw(thr, name_and_params):
-    name, params = name_and_params
-    distribution = 'uniform_integer'
     size = 1000
     seed = 123
 
-    dest = thr.array((params['words'], size),
-        numpy.uint32 if params['bitness'] == 32 else numpy.uint64)
-    rng = CBRNG(dest, 1, seed=seed, rng=name, rng_params=params, distribution=distribution)
-    rngc = rng.compile(thr)
+    bijection = test_bijection.bijection
+    keygen = KeyGenerator.create(bijection, seed=seed, reserve_id_space=False)
+    counters_ref = numpy.zeros((size, bijection.counter_words), bijection.dtype)
 
-    counters = create_counters(size, params)
-    counters_dev = thr.to_device(counters)
+    rng_kernel = thr.compile_static(
+        """
+        <%
+            ctype = dtypes.ctype(bijection.dtype)
+        %>
+        KERNEL void test(GLOBAL_MEM ${ctype} *dest, int ctr)
+        {
+            VIRTUAL_SKIP_THREADS;
+            const int idx = virtual_global_id(0);
 
-    rngc(counters_dev, dest, counters_dev)
-    dest_ref = rng_ref(0, size, name, seed, **params)
+            ${bijection.module}KEY key = ${keygen.module}key_from_int(idx);
+            ${bijection.module}COUNTER counter = ${bijection.module}make_counter_from_int(ctr);
+            ${bijection.module}COUNTER result = ${bijection.module}bijection(key, counter);
+
+            %for i in range(bijection.counter_words):
+            dest[idx * ${bijection.counter_words} + ${i}] = result.v[${i}];
+            %endfor
+        }
+        """,
+        'test', size,
+        render_kwds=dict(bijection=bijection, keygen=keygen))
+
+    dest = thr.array((size, bijection.counter_words), bijection.dtype)
+
+    rng_kernel(dest, numpy.int32(0))
+    dest_ref = test_bijection.reference(counters_ref, keygen.reference)
     assert diff_is_negligible(dest.get(), dest_ref)
 
-    rngc(counters_dev, dest, counters_dev)
-    dest_ref = rng_ref(1, size, name, seed, **params)
+    rng_kernel(dest, numpy.int32(1))
+    counters_ref[:,-1] += 1
+    dest_ref = test_bijection.reference(counters_ref, keygen.reference)
     assert diff_is_negligible(dest.get(), dest_ref)
 
 
-
-def check_distribution(thr, rng_name, rng_params,
-        distribution, distribution_params, dtype, reference):
+def check_kernel_sampler(thr, sampler, extent=None, mean=None, std=None):
 
     size = 10000
     batch = 100
     seed = 456
 
-    dest_dev = thr.array((batch, size), dtype)
-    rng = CBRNG(dest_dev, 1, seed=seed, rng=rng_name, rng_params=rng_params,
-        distribution=distribution, distribution_params=distribution_params)
-    rngc = rng.compile(thr)
+    bijection = sampler.bijection
+    keygen = KeyGenerator.create(bijection, seed=seed)
 
-    counters = create_counters(size, rng_params)
-    counters_dev = thr.to_device(counters)
+    rng_kernel = thr.compile_static(
+        """
+        KERNEL void test(GLOBAL_MEM ${ctype} *dest, int ctr_start)
+        {
+            VIRTUAL_SKIP_THREADS;
+            const int idx = virtual_global_id(0);
 
-    rngc(counters_dev, dest_dev, counters_dev)
-    dest = dest_dev.get()
+            ${bijection.module}KEY key = ${keygen.module}key_from_int(idx);
+            ${bijection.module}COUNTER ctr = ${bijection.module}make_counter_from_int(ctr_start);
+            ${bijection.module}STATE st = ${bijection.module}make_state(key, ctr);
 
-    extent = reference.get('extent', None)
-    mean = reference.get('mean', None)
-    std = reference.get('std', None)
+            ${sampler.module}RESULT res;
 
+            for(int j = 0; j < ${batch}; j++)
+            {
+                res = ${sampler.module}sample(&st);
+
+                %for i in range(sampler.randoms_per_call):
+                dest[j * ${size * sampler.randoms_per_call} + ${size * i} + idx] = res.v[${i}];
+                %endfor
+            }
+
+            ${bijection.module}COUNTER next_ctr = ${bijection.module}get_next_unused_counter(st);
+        }
+        """,
+        'test', size,
+        render_kwds=dict(
+            size=size, batch=batch, ctype=dtypes.ctype(sampler.dtype),
+            bijection=bijection, keygen=keygen, sampler=sampler))
+
+    dest = thr.array((batch, sampler.randoms_per_call, size), sampler.dtype)
+    rng_kernel(dest, numpy.int32(0))
+    dest = dest.get()
+
+    check_distribution(dest, extent=extent, mean=mean, std=std)
+
+
+def check_distribution(arr, extent=None, mean=None, std=None):
     if extent is not None:
-        assert dest.min() >= extent[0]
-        assert dest.max() <= extent[1]
+        assert arr.min() >= extent[0]
+        assert arr.max() <= extent[1]
 
     if mean is not None and std is not None:
         # expected mean and std of the mean of the sample array
         m_mean = mean
-        m_std = std / numpy.sqrt(batch * size)
+        m_std = std / numpy.sqrt(arr.size)
 
-        diff = abs(dest.mean() - mean)
+        diff = abs(arr.mean() - mean)
         assert diff < 5 * m_std # about 1e-6 chance of fail
 
     if std is not None:
         # expected mean and std of the variance of the sample array
         v_mean = std ** 2
-        v_std = numpy.sqrt(2. * std ** 4 / (batch * size - 1))
+        v_std = numpy.sqrt(2. * std ** 4 / (arr.size - 1))
 
-        diff = abs(dest.var() - v_mean)
+        diff = abs(arr.var() - v_mean)
         assert diff < 5 * v_std # about 1e-6 chance of fail
 
 
@@ -130,28 +180,25 @@ def uniform_mean_and_std(min, max):
 def test_32_to_64_bit(thr):
     extent = (0, 2**63-1)
     mean, std = uniform_discrete_mean_and_std(*extent)
-    check_distribution(thr,
-        'philox', dict(bitness=32, words=4),
-        'uniform_integer', dict(min=extent[0], max=extent[1] + 1), numpy.uint64,
-        dict(extent=extent, mean=mean, std=std))
+    bijection = philox(32, 4)
+    sampler = uniform_integer(bijection, numpy.uint64, extent[0], extent[1] + 1)
+    check_kernel_sampler(thr, sampler, extent=extent, mean=mean, std=std)
 
 
 def test_64_to_32_bit(thr):
     extent = (0, 2**31-1)
     mean, std = uniform_discrete_mean_and_std(*extent)
-    check_distribution(thr,
-        'philox', dict(bitness=64, words=4),
-        'uniform_integer', dict(min=extent[0], max=extent[1] + 1), numpy.uint32,
-        dict(extent=extent, mean=mean, std=std))
+    bijection = philox(64, 4)
+    sampler = uniform_integer(bijection, numpy.uint32, extent[0], extent[1] + 1)
+    check_kernel_sampler(thr, sampler, extent=extent, mean=mean, std=std)
 
 
 def test_uniform_integer(thr):
     extent = (-10, 98)
     mean, std = uniform_discrete_mean_and_std(*extent)
-    check_distribution(thr,
-        'philox', dict(bitness=64, words=4),
-        'uniform_integer', dict(min=extent[0], max=extent[1] + 1), numpy.int32,
-        dict(extent=extent, mean=mean, std=std))
+    bijection = philox(64, 4)
+    sampler = uniform_integer(bijection, numpy.int32, extent[0], extent[1] + 1)
+    check_kernel_sampler(thr, sampler, extent=extent, mean=mean, std=std)
 
 
 def test_uniform_float(thr_and_double):
@@ -159,20 +206,18 @@ def test_uniform_float(thr_and_double):
     dtype = numpy.float64 if double else numpy.float32
     extent = (-5, 7.7)
     mean, std = uniform_mean_and_std(*extent)
-    check_distribution(thr,
-        'philox', dict(bitness=64, words=4),
-        'uniform_float', dict(min=extent[0], max=extent[1]), dtype,
-        dict(extent=extent, mean=mean, std=std))
+    bijection = philox(64, 4)
+    sampler = uniform_float(bijection, dtype, extent[0], extent[1])
+    check_kernel_sampler(thr, sampler, extent=extent, mean=mean, std=std)
 
 
 def test_normal_bm(thr_and_double):
     thr, double = thr_and_double
     dtype = numpy.float64 if double else numpy.float32
     mean, std = -2, 10
-    check_distribution(thr,
-        'philox', dict(bitness=64, words=4),
-        'normal_bm', dict(mean=mean, std=std), dtype,
-        dict(mean=mean, std=std))
+    bijection = philox(64, 4)
+    sampler = normal_bm(bijection, dtype, mean=mean, std=std)
+    check_kernel_sampler(thr, sampler, mean=mean, std=std)
 
 
 def test_gamma(thr_and_double):
@@ -181,12 +226,44 @@ def test_gamma(thr_and_double):
     shape, scale = 3, 10
     mean = shape * scale
     std = numpy.sqrt(shape) * scale
-    check_distribution(thr,
-        'philox', dict(bitness=64, words=4),
-        'gamma', dict(mean=mean, std=std), dtype,
-        dict(shape=shape, scale=scale))
+    bijection = philox(64, 4)
+    sampler = gamma(bijection, dtype, shape=shape, scale=scale)
+    check_kernel_sampler(thr, sampler, mean=mean, std=std)
 
 
-@pytest.mark.perf
-def test_raw_perf():
-    pass
+def check_computation(thr, rng, extent=None, mean=None, std=None):
+    dest_dev = thr.empty_like(rng.parameter.randoms)
+    counters = rng.create_counters()
+    counters_dev = thr.to_device(counters)
+    rngc = rng.compile(thr)
+
+    rngc(counters_dev, dest_dev)
+    dest = dest_dev.get()
+    check_distribution(dest, extent=extent, mean=mean, std=std)
+
+
+def test_computation_general(thr_and_double):
+
+    size = 10000
+    batch = 101
+
+    thr, double = thr_and_double
+    dtype = numpy.float64 if double else numpy.float32
+    mean, std = -2, 10
+    bijection = threefry(64, 4)
+    sampler = normal_bm(bijection, dtype, mean=mean, std=std)
+
+    rng = CBRNG(Type(dtype, shape=(batch, size)), 1, sampler)
+    check_computation(thr, rng, mean=mean, std=std)
+
+
+def test_computation_convenience(thr):
+
+    size = 10000
+    batch = 101
+
+    extent = (0, 511)
+    mean, std = uniform_discrete_mean_and_std(*extent)
+    rng = CBRNG.uniform_integer(Type(numpy.int32, shape=(batch, size)), 1,
+        sampler_kwds=dict(low=extent[0], high=extent[1] + 1))
+    check_computation(thr, rng, extent=extent, mean=mean, std=std)